Module databricks_rust_catalog::data::permissions

Structs

ObjectOwnerResponse

PrivilegeAssignment

PrivilegeAssignmentsResponse

User

Enums

SecurableType

Enum representing various Unity Catalog (UC) securable objects Refer to: https://docs.databricks.com/en/data-governance/unity-catalog/manage-privileges/privileges.html

Functions

authenticate_user

Returns User Struct containing the princpal used for authentication.

can_read

Checks if a principal can read a given object from Unity Catalog API.

can_write

Checks if a principal can modify a given object from Unity Catalog API.

check_permissions 🔒

Checks if a principal has the specified permissions on a given object.

fetch_permissions 🔒

Fetches permissions for a given object from Unity Catalog API.

get_object_owner 🔒

Fetches the owner of a specified object from Unity Catalog API. Reference: /api/2.1/unity-catalog/{securable_type}/{full_name}